test: split CLI/Studio parity physics case into its own file for CI shard balance
cliStudioParity.test.ts ran 79s as one file. The P6 --include-physics
drop-test parity case (32s) moves to cliStudioParity.physics.test.ts;
the two kinematic-only parity cases stay. The stdout-capture plumbing
(captureValidateJson, sortedCodes) and fixture paths move unchanged to
cliStudioParityShared.ts.

No dedup possible: the three cases differ in fixture and/or
includePhysics. Test counts unchanged: 3 -> 2+1. Measured per-file:
23.0s + 25.6s.
