+++ tests/integration/physics-loop/cliStudioParity.test.ts
@@ -16,20 +16,19 @@
 // the canonical broken fixture. If a future PR re-opens the split
 // (e.g. wires the probe into only one surface), this test fails in
 // CI before the PR can merge.
+//
+// NOTE: companion file cliStudioParity.physics.test.ts was split out for
+// CI shard balance (per-file vitest sharding); it hosts the P6
+// --include-physics parity case. Shared plumbing: cliStudioParityShared.ts.
 
 import { describe, it, expect } from 'vitest';
-import { resolve } from 'node:path';
-import { runValidateCli } from '../../../src/agent/cli/commands/validate';
 import { reviewCadTool } from '../../../src/agent/mcp/tools/reviewCad';
-
-const VEC3_SPRING_BROKEN_FIXTURE = resolve(
-  __dirname,
-  '../../fixtures/mechanism/vec3-spring-broken.kcad.ts',
-);
-const CLEVIS_HINGE_REAL_FIXTURE = resolve(
-  __dirname,
-  '../../fixtures/mechanism/clevis-hinge-real.kcad.ts',
-);
+import {
+  VEC3_SPRING_BROKEN_FIXTURE,
+  CLEVIS_HINGE_REAL_FIXTURE,
+  captureValidateJson,
+  sortedCodes,
+} from './cliStudioParityShared';
 
 describe('CLI / Studio parity for mechanism truth', () => {
   it(
@@ -39,8 +38,8 @@ describe('CLI / Studio parity for mechanism truth', () => {
       // parse the structured mechanism field straight off stdout.
       // Physics OFF here — the parity check pins the kinematic-only
       // verdict (criteria 1-4) which both surfaces have shipped since
-      // P1; the dedicated physics-parity case below covers criteria
-      // 5+6 separately.
+      // P1; the dedicated physics-parity case in
+      // cliStudioParity.physics.test.ts covers criteria 5+6 separately.
       const cliResult = await captureValidateJson({
         file: VEC3_SPRING_BROKEN_FIXTURE,
         includeInterference: true,
@@ -103,80 +102,4 @@ describe('CLI / Studio parity for mechanism truth', () => {
     },
     180000,
   );
-
-  it(
-    'P6: clevis-hinge with --include-physics — CLI and Studio agree the bare hinge drops under gravity',
-    async () => {
-      // The clevis-hinge fixture passes the kinematic-only gate but, as
-      // a single revolute joint with no spring or actuator, it cannot
-      // pass the drop-test. Asserting both surfaces emit the same
-      // `mechanism.drops-on-release` failure pins the new criterion's
-      // wiring at parity.
-      const cliResult = await captureValidateJson({
-        file: CLEVIS_HINGE_REAL_FIXTURE,
-        includeInterference: true,
-        includePhysics: true,
-      });
-      const cliCodes = sortedCodes(cliResult.mechanismFailures ?? []);
-      expect(cliResult.mechanism).toBe('broken');
-      expect(cliCodes).toContain('mechanism.drops-on-release');
-
-      const review = await reviewCadTool({
-        file: CLEVIS_HINGE_REAL_FIXTURE,
-        includeInterference: true,
-        includePhysics: true,
-      });
-      const studioMechanism = (review as { mechanism?: string }).mechanism;
-      const studioFailures = (review as { mechanismFailures?: Array<{ code: string }> })
-        .mechanismFailures ?? [];
-      const studioCodes = sortedCodes(studioFailures);
-      expect(studioMechanism).toBe('broken');
-      expect(studioCodes).toEqual(cliCodes);
-    },
-    180000,
-  );
 });
-
-interface CapturedCliResult {
-  mechanism?: string;
-  mechanismFailures?: Array<{ code: string }>;
-}
-
-async function captureValidateJson(input: {
-  file: string;
-  includeInterference: boolean;
-  includePhysics: boolean;
-}): Promise<CapturedCliResult> {
-  // runValidateCli writes its JSON output to console.log. We intercept
-  // it so the test can read the structured mechanism field without
-  // shelling out.
-  const captured: string[] = [];
-  const originalLog = console.log;
-  console.log = (...args: unknown[]) => {
-    captured.push(args.map(String).join(' '));
-  };
-  try {
-    await runValidateCli({
-      file: input.file,
-      json: true,
-      includeInterference: input.includeInterference,
-      epsilon: 0.01,
-      physical: false,
-      includePhysics: input.includePhysics,
-    });
-  } finally {
-    console.log = originalLog;
-  }
-  const stdout = captured.join('\n');
-  // The JSON object should be the entire stdout (in --json mode there
-  // are no other writes).
-  const parsed = JSON.parse(stdout) as CapturedCliResult & {
-    mechanism?: string;
-    mechanismFailures?: Array<{ code: string }>;
-  };
-  return parsed;
-}
-
-function sortedCodes(failures: ReadonlyArray<{ code: string }>): string[] {
-  return failures.map((f) => f.code).sort();
-}
